is_prime lib
