+++ zad17.py
@@ -1,19 +1,7 @@
+from is_prime import is_prime
+
+
 def zad17(a,b):
-
-    def is_prime(n):
-        if n < 2:
-            return False
-        elif n == 2:
-            return True
-        elif n % 2 == 0:
-            return False
-
-        div = 3
-        while div * div <= n:
-            if n % div == 0:
-                return False
-            div += 2
-        return True
 
     def num_len(n):
         if n == 0:
